do not use /home/jovyan for config or installation
Some tweaks to the docker layer caching are made in the process...
overall files should be copied into the container as required so
that cache invalidation occurs as late as possible in the Dockerfile.

Instead of using the home directory for config, /etc files are
used to store config content for all users.

Permissions to files should be fixed in the docker layer that they
are created in to minimize layer size. The fix-permissions script
does this recursively with minimal modifications to the file tree.

Software installation should go in /opt instead of home