Fix issue with traefik issueing certificates with letsencrypt acme
The core of the issue was that for entrypoints a default certResolver
was not being specified. Additionally had to read some docs within
traefik https://doc.traefik.io/traefik/routing/entrypoints/#tls that
state that is a `tls = {}` section is specified within the crd
ingressroute the tls configuration is overwritten.

```
If a TLS section (i.e. any of its fields) is user-defined, then the
default configuration does not apply at all.
```

Related to #1015